Home.js, Profile.js deleted, Login.js, Register.js moved to folder Login and Loading.js introduced in Loading folder
