--- Newsaholic/src/containers/Home.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import { View, ImageBackground } from 'react-native';
-
-const Home = () => {
-  return (
-    <ImageBackground
-      
-    >
-      <Text>
-          HIIIIII
-      </Text>
-    </ImageBackground>
-  );
-};
-
-export default Home;
