--- Newsaholic/src/containers/Profile.js
@@ -1,37 +1,0 @@
-import React from 'react';
-// import styles from '../assets/styles';
-
-import {
-  ScrollView,
-  View,
-  Text,
-  ImageBackground,
-  TouchableOpacity
-} from 'react-native';
-// import ProfileItem from '../components/ProfileItem';
-// import Icon from '../components/Icon';
-// import Demo from '../assets/data/demo.js';
-
-const Profile = () => {
-  const {
-    age,
-    image,
-    info1,
-    info2,
-    info3,
-    info4,
-    location,
-    match,
-    name
-  } = Demo[7];
-
-  return (
-    <ImageBackground
-     
-    >
-        <Text>Hiii</Text>
-    </ImageBackground>
-  );
-};
-
-export default Profile;
