+++ Newsaholic/src/containers/screens/Loading/Loading.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import { Text, View, SafeAreaView } from 'react-native';
+
+export default class Loading extends React.Component {
+    render() {
+        return (
+            <View>
+                <SafeAreaView>
+                    <Text>
+                        Loading
+                    </Text>
+                </SafeAreaView>
+            </View >
+        )
+    }
+}
